requirements and train loop updated

+++ sub_ocr/postprocess/rec_postprocess.py
@@ -37,7 +37,7 @@ class BaseRecLabelDecode:
                 else:
                     conf_list.append(1)
             text = ''.join(char_list)
-            result_list.append((text, np.mean(conf_list) if conf_list else 0))
+            result_list.append((text, float(np.mean(conf_list)) if conf_list else 0))
         return result_list
 
     def get_ignored_tokens(self):
